requirements.txt
spacy
fasttext-wheel
pandas
numpy
scikit-learn
flask
joblib
pyspellchecker